creating a cleaner app.js file and creating components for sections of main app
